requirements.txt
openai
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
streamlit
tiktoken
pypdf
chromadb
pytest
transformers
jsonschema
python-docx
